refactor: package 오타 수정
