+++ src/main/java/racingcar/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racing;
+package racingcar;
 
 public interface RandomNumberGenerator {
